refactor: domain driver structure to function based structure
