--- internal/api/routes.go
@@ -1,39 +1,0 @@
-// Copyright (C) 2025 Nikash Labs
-//
-// This file is part of hishab.
-//
-// hishab is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// hishab is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with hishab.  If not, see <https://www.gnu.org/licenses/>.
-
-package api
-
-import (
-	"net/http"
-
-	"github.com/nikashlabs/hishab/internal/server/config"
-	"go.uber.org/zap"
-)
-
-func AddRoutes(
-	mux *http.ServeMux,
-	logger *zap.Logger,
-	config *config.Config,
-) {
-	mux.Handle("/", http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Write([]byte("Hi, Welcome Hishab Project\n"))
-	}))
-	// mux.Handle("/api/v1/comments", handleComments(logger, commentStore))
-	// mux.Handle("/api/v1/another", handleAnother(logger, anotherStore))
-	// mux.HandleFunc("/healthz", handleHealthz(logger))
-	// mux.Handle("/", http.NotFoundHandler())
-}
